feat!: moved algo tests in test/model
