+++ src/test/kotlin/model/MSFFinderTest.kt
@@ -1,7 +1,5 @@
-package viewmodel
+package model
 
-import model.MSFFinder
-import model.UndirectedGraph
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
