+++ src/test/kotlin/model/SCCCalculatorTest.kt
@@ -1,9 +1,6 @@
-package viewmodel
+package model
 
-import model.DirectedGraph
-import model.Vertex
 import androidx.compose.ui.graphics.Color
-import model.SCCCalculator
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import space.kscience.kmath.operations.IntRing
